change: Split CIB file into it's own module
As mentioned in #426, CIB file was already more than 3000 lines of code
and was turning into a deposit of all sorts of code for different
purposes.

This is a preliminary work towards adding a bit of order into the
CIB/bus-mapping crate.
